+++ src/_packages/utils.index.ts
@@ -46,3 +46,8 @@ export {
     trimEndOfCodeBlock,
     unwrapResult,
 };
+
+/**
+ * TODO: !!!! prune utils in 32
+ * TODO: [🧠] Maybe create some indipendent package like `markdown-tools` from both here exported and @private utilities
+ */