+++ CHANGELOG.md
@@ -77,12 +77,17 @@ Better execution report in markdown format
 
 ## 0.31.0 _(2024-02-XX)_
 
+Better execution reports
+
 -   Filter out voids in `executionReportJsonToString` <-[🎡]
 -   Add timing information to `ExecutionReportJson` <-[🎡]
 -   Add money cost information to `ExecutionReportJson` <-[🎡]
 -   Escape code blocks in markdown
 
 ## 0.32.0 _(2024-02-XX)_
+
+Iterating over parameters
+(!!! And maybe nullable parameters)
 
 -   Parameters can be both `string` and `Array<string>` <-(TODO: !!!!)
     -   `Array<string>` will itterate over all values <-(TODO: !!!!)